moved link generation back into each relevant resource; ensure that all self links have a matching streamStore: link
