+++ src/SqlStreamStore.HAL/Resources/AllStreamResource.cs
@@ -1,6 +1,7 @@
 ﻿namespace SqlStreamStore.HAL.Resources
 {
     using System;
+    using System.Collections.Generic;
     using System.Linq;
     using System.Net.Http;
     using System.Threading;
@@ -47,9 +48,9 @@
                         page.NextPosition,
                         page.IsEnd
                     })
-                    .AddLinks(Links.All.SelfFeed(operation))
-                    .AddLinks(Links.All.Navigation(page, operation))
-                    .AddLinks(Links.All.Feed(operation))
+                    .AddLinks(Links.Self(operation))
+                    .AddLinks(Links.Navigation(page, operation))
+                    .AddLinks(Links.Feed(operation))
                     .AddEmbeddedCollection(
                         Constants.Relations.Message,
                         streamMessages.Zip(
@@ -65,7 +66,7 @@
                                     payload,
                                     metadata = message.JsonMetadata
                                 })
-                                .AddLinks(Links.All.Self(message)))));
+                                .AddLinks(Links.Message.Self(message)))));
 
             if(operation.FromPositionInclusive == Position.End)
             {
@@ -77,6 +78,86 @@
             }
 
             return response;
+        }
+
+        private static class Links
+        {
+            public static Link Self(ReadAllStreamOperation operation)
+                => new Link(Constants.Relations.Self, operation.Self);
+
+            public static Link First(ReadAllStreamOperation operation)
+                => new Link(
+                    Constants.Relations.First,
+                    LinkFormatter.FormatForwardLink(
+                        Constants.Streams.All,
+                        operation.MaxCount,
+                        Position.Start,
+                        operation.EmbedPayload));
+
+            public static Link Last(ReadAllStreamOperation operation)
+                => new Link(
+                    Constants.Relations.Last,
+                    LinkFormatter.FormatBackwardLink(
+                        Constants.Streams.All,
+                        operation.MaxCount,
+                        Position.End,
+                        operation.EmbedPayload));
+
+            public static Link Last()
+                => new Link(
+                    Constants.Relations.Last,
+                    LinkFormatter.FormatBackwardLink(
+                        Constants.Streams.All,
+                        Constants.MaxCount,
+                        Position.End,
+                        false));
+
+            public static Link Feed(ReadAllStreamOperation operation)
+                => new Link(Constants.Relations.Feed, operation.Self);
+
+            public static Link Feed()
+                => new Link(Constants.Relations.Feed, Last().Href);
+
+            public static Link Previous(ReadAllPage page, ReadAllStreamOperation operation)
+                => new Link(
+                    Constants.Relations.Previous,
+                    LinkFormatter.FormatBackwardLink(
+                        Constants.Streams.All,
+                        operation.MaxCount,
+                        page.Messages.Min(m => m.Position) - 1,
+                        operation.EmbedPayload));
+
+            public static Link Next(ReadAllPage page, ReadAllStreamOperation operation)
+                => new Link(
+                    Constants.Relations.Next,
+                    LinkFormatter.FormatForwardLink(
+                        Constants.Streams.All,
+                        operation.MaxCount,
+                        page.Messages.Max(m => m.Position) + 1,
+                        operation.EmbedPayload));
+
+            public static IEnumerable<Link> Navigation(ReadAllPage page, ReadAllStreamOperation operation)
+            {
+                var first = First(operation);
+                var last = Last(operation);
+
+                yield return first;
+
+                if(operation.Self != first.Href && !page.IsEnd)
+                    yield return Previous(page, operation);
+
+                if(operation.Self != last.Href && !page.IsEnd)
+                    yield return Next(page, operation);
+
+                yield return last;
+            }
+
+            public static class Message
+            {
+                public static Link Self(StreamMessage message) => new Link(
+                    Constants.Relations.Self,
+                    $"streams/{message.StreamId}/{message.StreamVersion}");
+            }
         }
     }
 }
