+++ src/SqlStreamStore.HAL/Resources/StreamMessageResource.cs
@@ -1,6 +1,8 @@
 namespace SqlStreamStore.HAL.Resources
 {
     using System;
+    using System.Collections.Generic;
+    using System.Linq;
     using System.Net.Http;
     using System.Threading;
     using System.Threading.Tasks;
@@ -24,7 +26,7 @@ namespace SqlStreamStore.HAL.Resources
                 throw new ArgumentNullException(nameof(streamStore));
             _streamStore = streamStore;
         }
-        
+
         public async Task<Response> GetMessage(
             ReadStreamMessageByStreamVersionOperation operation,
             CancellationToken cancellationToken)
@@ -39,11 +41,12 @@ namespace SqlStreamStore.HAL.Resources
                             operation.StreamId,
                             operation.StreamVersion
                         })
-                        .AddLinks(Links.StreamMessage.Self(operation))
-                        .AddLinks(Links.StreamMessage.Navigation(operation)),
+                        .AddLinks(Links.Self(operation))
+                        .AddLinks(Links.Navigation(operation))
+                        .AddLinks(Links.Message(operation)),
                     404);
             }
-            
+
             if(operation.StreamVersion == StreamVersion.End)
             {
                 return new Response(new HALResponse(new object()), 307)
@@ -69,8 +72,9 @@ namespace SqlStreamStore.HAL.Resources
                         payload,
                         metadata = message.JsonMetadata
                     })
-                    .AddLinks(Links.StreamMessage.Self(operation))
-                    .AddLinks(Links.StreamMessage.Navigation(operation, message)));
+                    .AddLinks(Links.Self(operation))
+                    .AddLinks(Links.Navigation(operation, message))
+                    .AddLinks(Links.Message(operation)));
         }
 
         public async Task<Response> DeleteMessage(
@@ -78,9 +82,97 @@ namespace SqlStreamStore.HAL.Resources
             CancellationToken cancellationToken)
         {
             await operation.Invoke(_streamStore, cancellationToken);
-            
+
             return new Response(
                 new HALResponse(new HALModelConfig()));
+        }
+
+        private static class Links
+        {
+            public static Link Self(ReadStreamMessageByStreamVersionOperation operation) => new Link(
+                Constants.Relations.Self,
+                $"{operation.StreamVersion}");
+
+            public static Link Message(ReadStreamMessageByStreamVersionOperation operation) => new Link(
+                Constants.Relations.Message,
+                $"{operation.StreamVersion}");
+
+            public static Link First(ReadStreamPage page, ReadStreamOperation operation)
+                => new Link(
+                    Constants.Relations.First,
+                    LinkFormatter.FormatForwardLink(
+                        page.StreamId,
+                        operation.MaxCount,
+                        StreamVersion.Start,
+                        operation.EmbedPayload));
+
+            public static Link Previous(ReadStreamPage page, ReadStreamOperation operation)
+                => new Link(
+                    Constants.Relations.Previous,
+                    LinkFormatter.FormatBackwardLink(
+                        page.StreamId,
+                        operation.MaxCount,
+                        page.Messages.Min(m => m.StreamVersion) - 1,
+                        operation.EmbedPayload));
+
+            public static Link Next(ReadStreamPage page, ReadStreamOperation operation)
+                => new Link(
+                    Constants.Relations.Next,
+                    LinkFormatter.FormatForwardLink(
+                        page.StreamId,
+                        operation.MaxCount,
+                        page.Messages.Max(m => m.StreamVersion) + 1,
+                        operation.EmbedPayload));
+
+            public static Link Last(ReadStreamPage page, ReadStreamOperation operation)
+                => new Link(
+                    Constants.Relations.Last,
+                    LinkFormatter.FormatBackwardLink(
+                        page.StreamId,
+                        operation.MaxCount,
+                        StreamVersion.End,
+                        operation.EmbedPayload));
+
+            private static Link First() => new Link(Constants.Relations.First, $"{0}");
+
+            private static Link Previous(ReadStreamMessageByStreamVersionOperation operation) => new Link(
+                Constants.Relations.Previous,
+                $"{operation.StreamVersion - 1}");
+
+            private static Link Next(ReadStreamMessageByStreamVersionOperation operation) => new Link(
+                Constants.Relations.Next,
+                $"{operation.StreamVersion + 1}");
+
+            private static Link Last() => new Link(Constants.Relations.Last, $"{-1}");
+
+            private static Link Feed(ReadStreamMessageByStreamVersionOperation operation) => new Link(
+                Constants.Relations.Feed,
+                LinkFormatter.FormatBackwardLink(
+                    $"../{operation.StreamId}",
+                    Constants.MaxCount,
+                    StreamVersion.End,
+                    false));
+
+            public static IEnumerable<Link> Navigation(
+                ReadStreamMessageByStreamVersionOperation operation,
+                StreamMessage message = default(StreamMessage))
+            {
+                yield return First();
+
+                if(operation.StreamVersion > 0)
+                {
+                    yield return Previous(operation);
+                }
+
+                if(message.MessageId != default(Guid))
+                {
+                    yield return Next(operation);
+                }
+
+                yield return Last();
+
+                yield return Feed(operation);
+            }
         }
     }
 }
