+++ src/SqlStreamStore.HAL/Resources/StreamResource.cs
@@ -1,6 +1,7 @@
 namespace SqlStreamStore.HAL.Resources
 {
     using System;
+    using System.Collections.Generic;
     using System.Linq;
     using System.Net.Http;
     using System.Threading;
@@ -36,8 +37,8 @@ namespace SqlStreamStore.HAL.Resources
 
             var response = new Response(
                 new HALResponse(result)
-                    .AddLinks(Links.Stream.Self(operation))
-                    .AddLinks(Links.Stream.Feed(operation)),
+                    .AddLinks(Links.Self(operation))
+                    .AddLinks(Links.Feed(operation)),
                 result.CurrentVersion == 0
                     ? 201
                     : 200);
@@ -70,10 +71,10 @@ namespace SqlStreamStore.HAL.Resources
                         page.NextStreamVersion,
                         page.IsEnd
                     })
-                    .AddLinks(Links.Stream.Self(operation))
-                    .AddLinks(Links.Stream.Navigation(page, operation))
-                    .AddLinks(Links.Stream.Feed(page, operation))
-                    .AddLinks(Links.Stream.Metadata(operation))
+                    .AddLinks(Links.Self(operation))
+                    .AddLinks(Links.Navigation(page, operation))
+                    .AddLinks(Links.Feed(operation))
+                    .AddLinks(Links.Metadata(operation))
                     .AddEmbeddedCollection(
                         Constants.Relations.Message,
                         streamMessages.Zip(
@@ -89,7 +90,7 @@ namespace SqlStreamStore.HAL.Resources
                                     payload,
                                     metadata = message.JsonMetadata
                                 })
-                                .AddLinks(Links.StreamMessage.Self(message)))),
+                                .AddLinks(Links.Message.Self(message)))),
                 page.Status == PageReadStatus.StreamNotFound ? 404 : 200);
         }
 
@@ -98,6 +99,103 @@ namespace SqlStreamStore.HAL.Resources
             await operation.Invoke(_streamStore, cancellationToken);
 
             return new Response(new HALResponse(new object()));
+        }
+
+        private static class Links
+        {
+            public static Link First(ReadStreamPage page, ReadStreamOperation operation)
+                => new Link(
+                    Constants.Relations.First,
+                    LinkFormatter.FormatForwardLink(
+                        page.StreamId,
+                        operation.MaxCount,
+                        StreamVersion.Start,
+                        operation.EmbedPayload));
+
+            public static Link Previous(ReadStreamPage page, ReadStreamOperation operation)
+                => new Link(
+                    Constants.Relations.Previous,
+                    LinkFormatter.FormatBackwardLink(
+                        page.StreamId,
+                        operation.MaxCount,
+                        page.Messages.Min(m => m.StreamVersion) - 1,
+                        operation.EmbedPayload));
+
+            public static Link Next(ReadStreamPage page, ReadStreamOperation operation)
+                => new Link(
+                    Constants.Relations.Next,
+                    LinkFormatter.FormatForwardLink(
+                        page.StreamId,
+                        operation.MaxCount,
+                        page.Messages.Max(m => m.StreamVersion) + 1,
+                        operation.EmbedPayload));
+
+            public static Link Last(ReadStreamPage page, ReadStreamOperation operation)
+                => new Link(
+                    Constants.Relations.Last,
+                    LinkFormatter.FormatBackwardLink(
+                        page.StreamId,
+                        operation.MaxCount,
+                        StreamVersion.End,
+                        operation.EmbedPayload));
+
+            public static Link Self(ReadStreamOperation operation) => new Link(
+                Constants.Relations.Self,
+                operation.Self);
+
+            public static Link Self(AppendStreamOperation operation) => new Link(
+                Constants.Relations.Self,
+                $"{operation.StreamId}");
+
+            public static Link Feed(ReadStreamOperation operation)
+                => new Link(Constants.Relations.Feed, operation.Self);
+
+            public static Link Feed(ReadStreamMessageByStreamVersionOperation operation)
+                => new Link(
+                    Constants.Relations.Feed,
+                    LinkFormatter.FormatBackwardLink(
+                        operation.StreamId,
+                        Constants.MaxCount,
+                        StreamVersion.End,
+                        false));
+
+            public static Link Feed(AppendStreamOperation operation)
+                => new Link(
+                    Constants.Relations.Feed,
+                    LinkFormatter.FormatBackwardLink(
+                        operation.StreamId,
+                        Constants.MaxCount,
+                        StreamVersion.End,
+                        false));
+
+            public static Link Metadata(ReadStreamOperation operation)
+                => new Link(
+                    Constants.Relations.Metadata,
+                    $"{operation.StreamId}/metadata");
+
+            public static IEnumerable<Link> Navigation(ReadStreamPage page, ReadStreamOperation operation)
+            {
+                var first = First(page, operation);
+
+                var last = Last(page, operation);
+
+                yield return first;
+
+                if(operation.Self != first.Href && !page.IsEnd)
+                    yield return Previous(page, operation);
+
+                if(operation.Self != last.Href && !page.IsEnd)
+                    yield return Next(page, operation);
+
+                yield return last;
+            }
+
+            public static class Message
+            {
+                public static Link Self(StreamMessage message) => new Link(
+                    Constants.Relations.Self,
+                    $"{message.StreamId}/{message.StreamVersion}");
+            }
         }
     }
 }
